webserver/requirements.txt
Flask=1.1.1
python=3.7
pandas
tensorflow=1.15
transformers=2.5.1
pytorch=1.2.0
